Add spark conf for memory/cpu

+++ starfish-emr/src/main/java/org/datapunch/starfish/api/spark/ExecutorSpec.java
@@ -1,17 +1,17 @@
 package org.datapunch.starfish.api.spark;
 
 public class ExecutorSpec extends PodSpec {
-    private int instances;
+    private long instances;
 
     public ExecutorSpec() {
         this.instances = 1;
     }
 
-    public int getInstances() {
+    public long getInstances() {
         return instances;
     }
 
-    public void setInstances(int instances) {
+    public void setInstances(long instances) {
         this.instances = instances;
     }
 }
